Function executes a function given as a parameter on each element of an array

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -11,10 +11,10 @@
  */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	size_t i;
+	int i;
 
 	if (!array || !size || !action)
 		return;
-	for (i = 0; i < size; i++)
+	for (i = 0; i < (int)size; i++)
 		action(array[i]);
 }
